+++ ContentPatcher/Framework/Commands/Commands/DumpCommand.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using ContentPatcher.Framework.Conditions;
 using ContentPatcher.Framework.Patches;
-using Pathoschild.Stardew.Common.Utilities;
+using Pathoschild.Stardew.Common.Commands;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Commands.Commands