+++ ContentPatcher/Framework/Commands/Commands/InvalidateCommand.cs
@@ -1,3 +1,4 @@
+using Pathoschild.Stardew.Common.Commands;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Commands.Commands