+++ ContentPatcher/Framework/Commands/Commands/ReloadCommand.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Linq;
+using Pathoschild.Stardew.Common.Commands;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Commands.Commands