+++ ContentPatcher/Framework/Commands/Commands/UpdateCommand.cs
@@ -1,4 +1,5 @@
 using System;
+using Pathoschild.Stardew.Common.Commands;
 using StardewModdingAPI;
 
 namespace ContentPatcher.Framework.Commands.Commands